refactor: Separate CLI from app orchestration
Moves the core application setup logic from `src/cli.ts` to a new `src/app.ts` module. This improves separation of concerns by making the CLI a thin wrapper responsible only for argument parsing, while the new `app` module handles the composition of services like providers, tool executors, and the agent itself.

The monolithic `main` function has been broken down into smaller, more focused functions within `app.ts`, improving readability and maintainability.

The provider creation logic has been refactored from a large `switch` statement to a more scalable, data-driven approach using a map of initializers.
